Implement `LayerDataWithLogger` to add logger to `LayerData`.
Currently, to have loggers in the memory layer, we add them to the
`MemoryUsageLayerData` which extends the `LayerData`. To Add loggers in
the same way to the other layers, we need to apply the same changes to
them which leads to duplicate code. Since there are layers that don't
need loggers, we don't add the loggers to the `LayerData` class.

`LayerDataWithLogger` extends the `LayerData` and adds the CSV loggers to
it. The layers that require logger can use it instead of `LayerData`.

This PR also refactors the memory usage layer to use the new class.
